Function operators

Operators to augment the way a function behaves.

quantity_index() remaps price arguments into quantity argument (and vice versa) to turn a price index into a quantity index.

grouped() makes a function applicable to grouped data.

balanced() makes a function balance the removal of NAs across multiple input vectors.

Usage

quantity_index(f)

grouped(f, ...)

balanced(f, ...)

Arguments

f

A function.

...

Additional arguments to f that should not be treated as grouped / should not be balanced.

Value

quantity_index() returns a function like f, except that the role of prices/quantities is reversed.

grouped() returns a function like f with a new argument group. This accepts a factor to split all other arguments in f (except those specified in ...) into groups before applying f to each group and combining the results. It is similar to ave(), but more general.

balanced() returns a function like f, except that missing values are balanced across all inputs when na.rm = TRUE (except those specified in ...). This is like using complete.cases() on the inputs of f.

Note

Additional arguments passed to ... are evaluated, so the return function doesn't evaluate these arguments lazily. In most cases these additional arguments are parameters like elasticity or switches like na.rm, and eager evaluation prevents these from being changed in, e.g., a loop. Lazy evaluation can be had by passing an anonymous function that partials out these arguments.

See also

Examples

p1 <- price6[[3]]
p0 <- price6[[2]]
q1 <- quantity6[[3]]
q0 <- quantity6[[2]]

#---- Quantity index ----
# Remap argument names to be quantities rather than prices

quantity_index(laspeyres_index)(q1 = q1, q0 = q0, p0 = p0)
#> [1] 1.175887

laspeyres_index(p1 = q1, p0 = q0, q0 = p0)
#> [1] 1.175887

# Works with the index_weights() functions, too

quantity_index(index_weights("Laspeyres"))(q0 = q0, p0 = p0)
#> [1] 0.96 2.70 2.47 0.91 6.58 0.48

#---- Group ----
# Calculate Tornqvist weights for two groups

f <- factor(rep(letters[1:2], each = 3))
tornqvist_weights <- grouped(index_weights("Tornqvist"))
tornqvist_weights(p1, p0, q1, q0, group = f)
#> [1] 0.18247009 0.33481172 0.48271819 0.12865397 0.81833234 0.05301369

# Calculate a mean like ave(), but with weights

x <- 1:6
w <- c(1:5, NA)
grouped_mean <- grouped(geometric_mean, na.rm = TRUE)
grouped_mean(x, w, group = f)
#> [1] 2.182247 2.182247 2.182247 4.527921 4.527921 4.527921

# Redistribute weights

w1 <- c(2, 4)
w2 <- 1:6

harmonic_mean(mapply(harmonic_mean, split(x, f), split(w2, f)), w1)
#> [1] 3.333333

wr <- grouped(scale_weights)(w2, group = f) * w1[f]
harmonic_mean(x, wr)
#> [1] 3.333333

#---- Balance ----
# Balance missing values for a Fisher index

fisher <- balanced(fisher_index)
fisher(p1, p0, q1, replace(q0, 3, NA), na.rm = TRUE)
#> [1] 0.9015449
fisher_index(p1[-3], p0[-3], q1[-3], q0[-3])
#> [1] 0.9015449

# Operators can be combined, but some care may be needed

grouped(balanced(fisher_mean), na.rm = TRUE)(x, w, group = f)
#> [1] 1.954017 1.954017 1.954017 4.499657 4.499657 4.499657
balanced(grouped(fisher_mean))(x, w, group = f, na.rm = TRUE)
#> [1] 1.954017 1.954017 1.954017 4.499657 4.499657