GEKS index

Calculate an inter-temporal GEKS price index over a rolling window, as described in chapter 7 of Balk (2008), by Ivancic et al. (2011), and in chapter 10 of the CPI manual (2020).

Usage

## Function factory
geks(f)

## Special cases
tornqvist_geks(p, q, period, product, 
               window = nlevels(period), n = window - 1, na.rm = FALSE)

fisher_geks(p, q, period, product, 
            window = nlevels(period), n = window - 1, na.rm = FALSE)

Arguments

f

A price_index function that uses information on both base and current-period prices and quantities, and satisfies the time-reversal test. Usually a Törnqvist, Fisher, or Walsh index.

p

A numeric vector of prices, the same length as q.

q

A numeric vector of quantities, the same length as p.

period

A factor, or something that can be coerced into one, that gives the corresponding time period for each element in p and q. The ordering of time periods follows the levels of period to agree with cut().

product

A factor, or something that can be coerced into one, that gives the corresponding product identifier for each element in p and q.

window

The length of the rolling window. The default is a window that encompasses all periods in period. Values that are neither integers nor length 1 are silently truncated to a length 1 integer.

n

A number giving the length of the index series for each window, starting from the end of the window. For example, if there are 13 periods in window, setting n = 1 gives the index for period 13. The default gives an index for each period in window. Values that are neither integers nor length 1 are silently truncated to a length 1 integer.

na.rm

Should missing values for p and q be removed when calculating the index? By default missing values will return a missing value for the index.

Value

geks() returns a function:

function(p, q, period, product, 
         window = nlevels(period), n = window - 1, na.rm = FALSE){...}

This calculates a period-over-period GEKS index with the desired index-number formula, returning a list for each window with a named-numeric vector of index values.

tornqvist_geks() and fisher_geks() both return a list with a named numeric vector giving the value of the respective period-over-period GEKS index for each window.

Note

Like back_period, if multiple prices correspond to a period-product pair, then the back price at a point in time is always the first price for that product in the previous period. Unlike a bilateral index, however, duplicated period-product pairs can have more subtle implications for a multilateral index.

References

Balk, B. M. (2008). Price and Quantity Index Numbers. Cambridge University Press.

ILO, IMF, OECD, Eurostat, UN, and World Bank. (2020). Consumer Price Index Manual: Theory and Practice. International Monetary Fund.

Ivancic, L., Diewert, W. E., and Fox, K. J. (2011). Scanner data, time aggregation and the construction of price indexes. Journal of Econometrics, 161(1): 24--35.

See also

Examples

price <- 1:6
quantity <- 6:1
period <- rep(1:3, 2)
product <- rep(letters[1:2], each = 3)

tornqvist_geks(price, quantity, period, product)
#> [[1]]
#>        2        3 
#> 1.530869 1.376227 
#> 

tornqvist_geks(price, quantity, period, product, window = 2)
#> [[1]]
#>        2 
#> 1.520408 
#> 
#> [[2]]
#>        3 
#> 1.366822 
#> 

# Missing data

quantity[2] <- NA

# Use all non-missing data

fisher_geks(price, quantity, period, product, na.rm = TRUE)
#> [[1]]
#>        2        3 
#> 1.455853 1.370478 
#> 

# Remove records with any missing data

fg <- geks(balanced(fisher_index))
fg(price, quantity, period, product, na.rm = TRUE)
#> [[1]]
#>        2        3 
#> 1.403078 1.346954 
#>